requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
cffi==1.15.1
cryptography==39.0.2
Django==4.1.7
django-cors-headers==3.14.0
django-jsonfield==1.4.1
django-redis==5.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7.1
redis==4.5.4
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
uuid==1.30
channels==4.0.0
channels-redis==4.1.0